secure-file-sharing/requirements.txt
Flask==2.2.5
pycryptodome==3.19.0
python-dotenv==1.0.0
